working but if blacklist removed, blacklisted ws close if trying to send message
